requirements.txt
torch
torchvision
timm
grad-cam
shap
numpy
matplotlib
streamlit
scikit-learn
Pillow